Update README.md
Hey- I noticed I didn't have a space between the ## and the 'add' in my header for the joke.  I also just changed the code directly on the README in git and wanted to see what happened.  I know this is probably not the best practice so I figured I would give it a go now instead of possibly learning the hard way down the road!